+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/domain/models/AccessToken.kt
@@ -1,4 +1,4 @@
-package ru.appdevelopers.githubclient.models
+package ru.appdevelopers.githubclient.domain.models
 
 open class AccessToken(
     open val accessToken: String? = null,
