+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/domain/models/ErrorEvent.kt
@@ -1,4 +1,4 @@
-package ru.appdevelopers.githubclient.models
+package ru.appdevelopers.githubclient.domain.models
 
 sealed class AuthErrorResponse
 class GoogleAuthErrorResponse<T>(val data: T) : AuthErrorResponse()
