+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/domain/models/User.kt
@@ -1,3 +1,3 @@
-package ru.appdevelopers.githubclient.models
+package ru.appdevelopers.githubclient.domain.models
 
 data class User(val type: AuthType, val userName: String? = null)
