+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/ui/auth/AuthFragment.kt
@@ -2,7 +2,6 @@ package ru.appdevelopers.githubclient.ui.auth
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -10,9 +9,11 @@ import kotlinx.android.synthetic.main.auth_fragment.*
 import moxy.presenter.InjectPresenter
 import ru.appdevelopers.githubclient.R
 import moxy.presenter.ProvidePresenter
+import ru.appdevelopers.githubclient.domain.models.AccessToken
+import ru.appdevelopers.githubclient.domain.models.AuthErrorResponse
+import ru.appdevelopers.githubclient.domain.models.GitHubAuthErrorResponse
+import ru.appdevelopers.githubclient.domain.models.GoogleAuthErrorResponse
 import ru.appdevelopers.githubclient.di.DIConfig
-import ru.appdevelopers.githubclient.googleAuth.GoogleAccessToken
-import ru.appdevelopers.githubclient.models.*
 import ru.appdevelopers.githubclient.ui.base.BaseFragment
 import toothpick.Toothpick
