+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/ui/base/BaseFragment.kt
@@ -2,7 +2,7 @@ package ru.appdevelopers.githubclient.ui.base
 
 import android.widget.Toast
 import moxy.MvpAppCompatFragment
-import ru.appdevelopers.githubclient.models.IBaseCallback
+import ru.appdevelopers.githubclient.ui.IBaseCallback
 
 abstract class BaseFragment: MvpAppCompatFragment(), IBaseCallback {
     override fun showProgress() {
