+++ GitHubClient/app/src/main/java/ru/appdevelopers/githubclient/ui/base/BasePresenter.kt
@@ -3,7 +3,7 @@ package ru.appdevelopers.githubclient.ui.base
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import moxy.MvpPresenter
-import ru.appdevelopers.githubclient.models.IBaseCallback
+import ru.appdevelopers.githubclient.ui.IBaseCallback
 
 abstract class BasePresenter<T : IBaseCallback> : MvpPresenter<T>()  {
